tooling/kit: kit now uses a loopback channel for managing internal events such as pushing status updates
This resolves a bug which cropped up with the remote watcher which uses a hidden channel for actual dispatching due to type system problems.  Effectively assumptions were made a channel with a specific type will always be used.  Not actually the case and the design needs to be revisisted onces more info + expierence is available.